Fix two layouts in the New loop dialog the templates work broke
Both from the design-fidelity pass, and both visible in one screenshot.

**The applied strip was a scrambled sentence.** The shape sentence, the `Undo
the shape` button and the clause after it sat in an `HStack`, so when the
sentence wrapped, its second line went *under* the button and the words after
it — what you read was "…and sets Undo the shape to keep just the prompt. a
cadence." It is now one `Text` carrying an inline link, which flows with the
words around it; the link is intercepted by the strip's own `OpenURLAction`.
The type name is emphasised, as the canvas draws it.

**The footer had no room.** The save notice sat between the Spacer and the
primary button, and there is not enough width for both: the path truncated to
"Saved…op.md", "Put it in the project instead" wrapped onto three lines, and
`Create timed loop` itself wrapped onto two. A dialog's primary action must
never wrap. The notice now has a full-width row of its own above the buttons,
the button is `lineLimit(1)` and fixed-size, and the disabled-reason text
truncates rather than squeezing it.

Verified by rendering the dialog headless (`NSHostingView` → PNG) rather than
by reading the code, which is how both of these got through in the first place.

`NodeDraftForm` went over swiftlint's type-body limit with the fix, so the
template chrome — the applied strip and the footer — moved to
`NodeDraftTemplateBar.swift`, the way `ProjectFeature`'s template verbs already
live beside it.

1482 tests pass, swiftlint 0 errors, swift-format clean.
